\name{multiCol-package}
\alias{multiCol-package}
\alias{multiCol}
\docType{package}
\title{
\packageTitle{multiCol}
}
\description{
\packageDescription{multiCol}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{multiCol}
\packageIndices{multiCol}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{multiCol}

Maintainer: \packageMaintainer{multiCol}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
